When the replay is changed, there is a callback done to the UI frame `OnReplayChange` handler.This handler will perform the special case where the state is `COMPLETED`, then we set it back to `STAGING` and call `:Reset()`.
The UI will also be hidden if not in a keystone related instance (i.e. state is `NONE`).

This makes it so the completed run data is stickies, until a new key is started, or a new instance is entered, or the replay is changed. So unless that happens, the player can exit and re-enter the same instance and the UI will show the previous run information.

This data is only in the UI and not saved in SV, so a /reload will always force the state back to default behavior. It can only recover by using live data, so entering a instance with an ongoing key will force the data to update to match what is happening.