Changed BossFrameMixin Setup method takes in order in addition to index and bossID.
This will help us to visually re-order boss frames later, but keep the original index intact.

Added the handler OnBossKill, just like with OnReplayChanged.

This one will help us control if boss frames require to be re-ordered, by running the SetUIBosses method.

This method has also been optimized to only release the pool, if there are changes to the boss data.

As long the boss frames are the same and in the same order, we skip releasing them, just to re-add them back.

This way we only do the full update when the frames are dirty and require cleaning.

They are considered dirty if either the pool count is different than the sortedBosses count, or if the order has changed.