Ensured that when the live data provider is asked for data, it only responds with the cached data if the state is not currently `PLAYING`.
Ensured that when `SaveLiveSummary` is called that the state is set to `COMPLETED` otherwise it will log a incomplete run.

Added a simple message when a run is logged to the disk, wording isn't perfect, can probably be skipped to not confuse the player with the actual website logging the run or something.

Fixed bug with `sortedBosses` logic where if the run is not `PLAYING` the live data is missing, as expected, but then we get `nil` errors. The fix is to rely on the replay data, and if we're live, then rely on the live data. This way the UI is consistent in how bosses are sorted.

The `OnEvent` handler will ensure it tries to call `SaveLiveSummary` only when it transitions from `PLAYING` state into `COMPLETED` state. This should eliminate weird cases of entering or leaving or reloading the UI during an active run, sometimes triggering it to think it completed.

Added `TrimHistoryFromSV` function in the replay module to remove entries older than 24 hours to avoid infinite growth. This happens when the module loads (or at login).