Added encounter start/end tracking. This helps us identify when we are in combat with an encounter or not, so we can start and stop the combat timer appropriately. The state is tracked in the `ActiveEncounters` table.
Adjusted `UnpackReplayEvent` so it handles toggling of combat, and setting and unsetting of the new `combatStart` property on replay bosses.

The UI uses the kill order instead of the boss index. Bosses which are killed first appear on top of the list, and the live boss kills always have priority in how the list is rendered.

Added combat timers for live and replay bosses, so it counts up as the parties are in combat. It is in green to signify it is active, as it turns white when it's set and finalized.

Added combat icons for both live and replay, so no longer just the replay side with an icon.

The live data provider will properly assign the order and id for the live bosses, as we have the data we need to figure that information out.

The live data provider also got the same treatment as the replay provider, where the combat and combatStart are set accordingly. Including the pull counter, though not yet used anywhere.

Finally fixed the bug with wrongfully color coded trash %. It was due to `min(replayTrash, totalTrash) - liveTrash` previously using `100` instead of `totalTrash`.

Overhauled the sorting code of `sortedBosses` and it should now be final and done.

Slight adjustment of state setting, we only set the COMPLETED state when transitioning from PLAYING, this way we avoid the situation where we go from NONE to COMPLETED, even if the run never completed. This makes it so the UI keeps the data, as long you stay inside, if you leave, re-enter, it resets to STAGING mode.