Unified all time conversion to rely on `ConvertMillisecondsToSeconds` to avoid ambigous conversions, some used ceil, some used floor, it made a bit of inconsistencies. The new method rounds, so 1.49 becomes 1, 1.5 becomes 2, etc.
Adjusted the parameters of `SecondsToTimeText` and replaced all occurences of either `util:SecondsToTimeText` or `SecondsToClock` to instead rely on this function. It supports the `splitStyle` argument to specify how we want the timer to look.

Adjusted the ongoing boss encounter timers will display as yellow, and the split will be in parenthesis.

Adjusted the logic behind `SetUITitle` so that if the key is running or completed, only then do we update and set the UI to display the level and affix on the live side, otherwise we would empty it for other stages like the staging phase.

Adjusted the way the liveTimer is calculated by substracting the deaths, as they are applied by the game itself. It does make the timer act a bit janky, as the event for the deaths and the timer update might happen with a tiny bit of gap of each other.