Boss combat detection should work as expected, even if changing the replay mid-run.
Added `deathsBeforeOvertime` property to the replay summary, where if deaths happen after the timelimit, they will register the true death amount on this properly. This is needed to counter the timer, where penalties are included in-time, but out of time they are not.

The replay `GetSummary` will ensure to finalize the combat state on bosses to cleanup if we process the last event in the replay.
Moved a lot of helper functions in their own places to make it easier to use where needed:
`:GetCurrentTimeLimit()` - to help retrieve the time limit from the replay dungeon or the data on the frame itself
`:GetCurrentDeaths()` - to help retrieve the deaths from live and replay, both with and without the penalty`:GetKeystoneTime()` - updated to return the real time elapsed (without penalties, unless requested)
`:RefreshWorldElapsedTimeState()` - updates the internal timer state with the server timer
Everywhere `.timer` is used is replaced with `:GetKeystoneTimeMS()` to combat ambiguous situations.

The `:OnUpdate()` handler will try to detect hickup frame drops, and refresh the internal timer using `RefreshWorldElapsedTimeState`.

The `:Update()` code will now properly handle death penalties, as it adds them on top of the true keystone timer.

When the key starts to play we also do a safety reset to cleanse any leftover from the "last snapshot" view during staging.