Minor changes to documentation as the language server improved and now detects required and optional properties.
Updated the replay icons with newer icons, and added icons for combat and route swap indicators.

Updated the replay boss sorting so that the order always matches the replay, and we simply match our kills in order, against the replay kills, so bosses might get out of sync, but that is ok, we indicate it with the route swap icon and when you hover you see the live vs replay boss names for clarity.

Slightly moved the main boss combat indicators to the side so they don't almost overlap the boss count text.