Attempt to fix issue with race conditions where the old objective tracker data would pollute the new tracker.
Avoiding this by encapsulating the live boss replay table so that when `encounter` is being requested, it will fetch the up-to-date encounter table for the boss.

This ensures the UI properly updates with the valid boss information.

Tried to mitigate nil issues by fallback to index if order is not set. Usually both are equal, except if the route doesn't follow the boss order, then the order is used to sort, and it should be present unless something is wrong.

Added back more missing isDirty check situations.