Added replay option to set timing method for the splits.
Default timing is BOSS combat duration comparisons, the other one is based on the defeat timer relative to dungeon time.

Fixed issue where route swap and combat indicator overlapped, the priority is to show combat indicator first, then route swap icon so they don't overlap.

Changed a lot on how the sorting of the UI works and how the encounter detection works, so it needs a bit more testing due to these changes.