`SecondsToTimeTextCompared` should properly compare both delta to see which one is ahead or not.
Calculating the `comparisonDelta` will properly avoid pre-calculations, as this caused the coloring step to fail and always show red text. Now, it will compare both this and the `delta` and color the text accordingly.

Removed all `debugMode` specific code, and instead added two simple `StartDebug` and `StopDebug` methods on the frame that can be used with debug enabled, to fake play and complete a run, just to see how the UI works. It's not 1:1 to reality and it's a bit janky due to the fact it doesn't have live API data to feed it, but it's a good enough approximation for testing some stuff.