Removed unused argument from `SetUITrash` and `SetUIDeaths` functions.
Ensured we store `replayCompletedTimer` the same way we do the other ms to seconds conversion timers by using `ConvertMillisecondsToSeconds`.

Updated SetUITimer so that it considers `replayCompletedTimer` if it exists when doing the coloring of the timer text so it doesn't get confused with the replay end time from the replay data.