Tried to improve replay UI anchoring by detecting if the objective frame is where it should be, and if not to fallback on the parent container and anchor next to it instead.
Unified server checking in `util:IsOnRetailRealm()` so that it is simple to evaluate if we are on live servers, or a test client or realm.

Fixed an issue that resulted in lua errors if the server region was not detected. We ensure to fallback to US region just to be on the safe side. This should never happen, but we will be alerted if it does by the built-in print message for new servers.