Since the dungeon file contains expansion dungeons the concept had to be added to the code.
Both season and expansion dungeons are placed in a `ALL_DUNGEONS` table, and each dungeon has the new property `type` assigned in case we wish to know if it's `SEASON` or `EXPANSION`.

The `ns:GetDungeonData()` was expanded to return `DUNGEONS`, `EXPANSION_DUNGEONS`, `ALL_DUNGEONS` instead of just the dungeons like before. This makes it possible to read additional args for either expansion specific table, or just all dungeons table.

The `util` module has had multiple "GetDungeon" prefixed functions updated to rely on all dungeons, this way we ensure that all modules can retrieve the dungeon data, even if it's not seasonal, and instead we must then filter or handle the differences. Going over, it seems most are fine as-is after the change.

The auto combat log module was also changed to use all dungeons when calling `getLowestMapIdForInstances`.

Modernized the old export dungeon data tools.